--- cmd/root.go
@@ -1,43 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package cmd
-
-import (
-	"os"
-
-	"github.com/apache/servicecomb-service-center/pkg/log"
-	"github.com/spf13/cobra"
-)
-
-var rootCmd = &cobra.Command{
-	Use:   "service-center",
-	Short: "service-center is a data centers",
-}
-
-func Execute() error {
-	if len(os.Args) < 2 || os.Args[1] != syncerCmd.Use {
-		runServiceCenter()
-		return nil
-	}
-
-	err := rootCmd.Execute()
-	if err != nil {
-		log.Errorf(err, "root command execute failed")
-	}
-	return err
-
-}
